Split up API test protocols into separate files
Protocol tests have become more complex since they
were first implemented. This splits NFS and SMB into
separate files.
